PROJECT RED!! v4
This is the start I'm Leaving it here as I will finish this one just want people to know it is being worked on.
Some Stuff Is pulled from RP2 as its a dead mod and and I have remade a few things and Some things will have to be done over completely.