ExtraTic Complete
I found the cause of the alpha transparency loss when using Justin_a_'s
script. If the overlay is completely desaturated, both the top and
bottom have no transparency and the image becomes greyscale. Upon
exporting the image, the file is compressed down to a bit depth of
eight. This bit depth does not support an alpha channel, and thus
Imagemagick occasionally uses the TRNs png chunk.
Minecraft does not support the TRNs chunk because it has a limited
implementation, so the background becomes black. That's why random
textures from the script had mysterious black backgrounds ingame. Adding
a miniscule amount of color to the overlay avoided the more compressed
bit depth and usage of the TRNs chunk altogether.