Carpenter's Blocks Update & Expansion
All Carpenter's Blocks textures now included. The wood texture is based
on the Pine Planks texture from Biomes O' Plenty support, and is
intended to look like unfinished wood. I think it works pretty well; the
doors look a lot like the ones I see at Home Depot. :-J

The existing hammer and chisel are more realistic than mine, but mine
are closer to the original textures. Put mine in alts, along with the
matching blanket for the hammer.

Cleaned up readme files that are better left in the main mod, added a
single Readme to explain the added structures and how to use the extra
blanket option. Renamed old alts to fit the current filename scheme.

I wasn't sure quite what the artistic goal was for a couple of the
patterns, so I just upsized them. If anyone else can tell what they are
supposed to be/look like, feel free to update them. I will keep looking
at them to see if I can figure them out. If they are better left out
entirely, that's fine.

Some of the quilt/blanket patterns still aren't great, but the color
combinations are more likely to get used than the originals. The
speckles are supposed to simulate the small patterns you usually see on
quilts (that's what it came out looking like when I shrank a picture of
a real quilt for personal use), but it's not as effective as I had
hoped.  Again, will be revisiting those, feedback is welcome.

Remembered to delete the random stackdump file this time. :-)

-NocturneBlock