Removal / Edits
- Removed textures which don't match the "Soartex Fanver" feel and that
don't suit the mod at all. This is nothing against the person who made
the textures but we need to stick to the fine line of the mod's original
texture and fitting to the Fanver feel.
- Moved some textures to alt folders as again these don't fit the
category of the Soartex Pack. I will design a better system to handle
alt textures.
- Some fixes/edits